Rewritten Quickterms to form a core language.
Due to these changes, the Quickterm data type represents now a programming
language. The predicates form computable actions which are applied on an input
program. They are used to decide which path of the Quickterm-program will be
taken. The resulting TermAction forms the actual executing code.

By that Quickterm is a command line argument parser interpreter. Whew, what
a name...

The code for Quickterm predicates can be formulated with simple Haskell
lambdas and form by that an EDSL. The Help-type is used for generation of help
messages. Due to function composition it is possible to form an EDSL for
manipulation of help messages. This makes it possible to adjust indentation
like seen in 'defaultWs' and do various other things. I could think of an EDSL
for composition of TermActions as the use of 'const' gives slight inside about
parameter skipping. A conversion EDSL could be used for easy serialization of
data.

At runtime the interpreter (implemented in 'runQuickterm') all predicates will
be applied on the argument list. The result of this application is used to
calculate a penalty. E.g. you can filter for (penalty==0) to find full matched
TermActions. Multiple results mean ambiguous predicates.

Due to penalty returning, it is now possible to calculate the most presumable
command line for error correction for further help for the user of the command
line application.